Refactor of helper
